# modern-orchestrall/docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Orchestrall AI Platform API
  description: |
    # Orchestrall AI Platform API
    
    A comprehensive multi-tenant AI orchestration platform that provides:
    - Universal CRUD operations for all entities
    - Real-time WebSocket updates
    - Plugin-based architecture
    - Multi-tenant support
    - Advanced analytics and monitoring
    
    ## Authentication
    
    The API uses JWT (JSON Web Token) for authentication. Include the token in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API requests are rate limited per endpoint:
    - General endpoints: 30 requests per minute
    - API endpoints: 10 requests per second
    - Authentication endpoints: 1 request per second
    
    ## Multi-tenancy
    
    All operations are automatically scoped to the authenticated user's organization.
    
  version: 2.0.0
  contact:
    name: Orchestrall AI Support
    email: support@orchestrall.ai
    url: https://orchestrall.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://orchestrall.ai/terms
servers:
  - url: https://api.orchestrall.ai/v2
    description: Production server
  - url: https://staging-api.orchestrall.ai/v2
    description: Staging server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@orchestrall.com
                password:
                  type: string
                  format: password
                  example: admin123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
                  expiresIn:
                    type: number
                    example: 3600
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Get a new JWT token using refresh token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  expiresIn:
                    type: number
                    example: 3600
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Entity Management Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "admin@orchestrall.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        role:
          type: string
          enum: [USER, ADMIN, MANAGER]
          example: "ADMIN"
        isActive:
          type: boolean
          example: true
        organizationId:
          type: string
          example: "org-123"
        createdAt:
          type: string
          format: date-time
          example: "2025-01-20T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-20T10:30:00Z"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        message:
          type: string
          example: "The request could not be processed"
        code:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: "2025-01-20T10:30:00Z"
        path:
          type: string
          example: "/api/users"
...